requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
japanize-matplotlib>=1.1.3
fastapi>=0.110
uvicorn>=0.29
xgboost>=2.0
lightgbm>=4.0
scikit-learn>=1.3
beautifulsoup4>=4.12
requests>=2.31
lxml>=5.0
streamlit>=1.30
plotly>=5.18
joblib>=1.3
wandb>=0.16
google-cloud-bigquery>=3.0
pyarrow>=14.0
db-dtypes>=1.0
